refactor: split apps router into domain-focused sub-routers (#2375)
